requirements.txt
Django==3.0.6
djangorestframework==3.11.0
freezegun==0.3.15
pytest==5.4.2
redis==3.5.2